aio: vmap entire aio_ring instead of kmapping each page
The aio_ring mechanism relies on two assumptions:
  1. The size of the header of an aio_ring is equal to the size of an
     io_event, i.e. sizeof(aio_ring) equals sizeof(io_event)
  2. There is no leftover space at the end of a page when populating it
     with io_events, i.e. PAGE_SIZE is a multiple of sizeof(io_event)

These assumptions are not met in PCuABI. This change removes these
assumptions, making the aio_ring mechanism more flexible and robust.

Mapping pages independently using kmap doesn't create a contiguous
virtual space. As some io_event structs could span across different
pages, all the pages necessary to access the events' data need to be
mapped to a contiguous virtual address space. For this reason, the ring
pages are vmapped at the context initialisation step, and the mapping is
cached in the context.

In addition, this change improves the performance of the aio_ring
mechanism. Previously, io_events were copied from the shared ring to
userspace using multiple copy_to_user calls. This change batches the
copying of io_events into a maximum of two copy_to_user calls.

The changes are similar to the approach taken in [0].

[0] https://lore.kernel.org/all/1349764760-21093-4-git-send-email-koverstreet@google.com

Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>