arm64: signal: Fix stale CSP read
When Morello support is enabled, there is no guarantee that the view
of a task's X and C registers (e.g. regs->regs and regs->cregs) is
coherent (see also the note on "write coalescing" in
Documentation/arm64/morello.rst). X/C merging always happens when
returning to userspace, but the value of a task's X registers may be
updated without the task being scheduled. This may happen due to
ptrace(PTRACE_SETREGSET), but also, crucially, when multiple pending
signals are delivered at the same time (typically when the task
unblocks them).

All capability registers are read in preserve_morello_context(), in
order to store them in the signal frame, and
morello_merge_cap_regs() is called beforehand to update the view of
capability registers.

However, there was an oversight in one of the initial Morello
support commits ("arm64: morello: Context-switch Restricted
registers"): regs->csp is also read in signal_sp(), which is called
before preserve_morello_context(). signal_sp() may therefore read a
stale SP value.

signal_sp() has changed several times since that commit, especially
to add PCuABI support. The last commit ("arm64: signal: Fix
signal_sp() in compat64") made the issue resurface, as regs->csp is
now used in compat64 too.

The error condition is indeed difficult to hit in PCuABI, as it
requires using ptrace with a very precise timing. On the other hand,
it can be reliably triggered in compat64 by delivering two signals
at the same time: regs->sp is set in setup_return() when the first
signal frame is set up, leaving regs->csp unchanged, and signal_sp()
will then read a stale SP value through regs->csp to set up the
second signal frame.

This situation is precisely what the sigpending02 LTP test creates,
allowing the bug to be finally caught.

To avoid any further issue, the call to morello_merge_cap_regs() is
therefore moved to the very beginning of setup_rt_frame(), before
any register is read.

Note that signal_sp() is also called from the rt_sigreturn handler,
but this is not a concern as the register view is up-to-date at the
point where a syscall handler is called.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>