arm64: futex: Enable capability-based uaccess
When working with PCuABI we need to ensure that we receive capabilities
from user space instead of pointers and perform the appropriate memory
accesses.

This commit updates the arm64-specific code for the futex module so that
the atomic load and store work with capabilities, hence enabling PCuABI.
This commit therefore reverts "arm64: futex: Access user memory via raw
64-bit pointers".

Because the instructions used to perform the atomic operations, ldxr and
stlxr, only support capabilities in C64 mode, this commit therefore
brings C64/A64 mode switching to the inline assembly.

Whenever a load/store faults, the exception raised will be handled
by ex_handler_uaccess_err_zero(), which will correctly set -EFAULT
as error and return execution past the loop. Exceptions are always
handled in A64, hence the mode is automatically switched back from
C64. The A64/C64 mode is saved in SPSR_ELx.C64 on exception entry,
thus C64 is restored upon return. What is more relevant in this case
is that the exception correctly returns to the fixup machinery
(label 3), which is when the switch back to A64 happens.

Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>