futex: Handle capability-based robust list entries
A robust list is implemented in the user space, but the kernel holds the
right to walk it. Given that a robust list may implement the priority
inheritance method, the approach chosen to inform the kernel of this
feature is to encode the LSB of the uaddr of a list entry to walk from.

In a PCuABI environment the userspace pointer is not a raw address but a
capability. So that we can clear the encoded priority-inheritance flag,
the helper USER_PTR_ALIGN_DOWN is used without resorting to
CHERI-specific code.

Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>