futex: Add explicit capability checking TODOs
Within the futex module, there are some cases in which a user address is
handled. When working with PCuABI this means that the capability is
discarded, and no checks can be performed by the hardware.

This commit adds TODOs whenever explicit capability checks need to be
performed, right before the raw pointer is extracted.

Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>