kselftests/arm64: morello: mmap: Add mmap() reservation testcases
Reservations are contiguous ranges of virtual addresses that exactly
match the bounds of an owning capability. When an owning capability is
passed to a syscall, its bounds are first verified against the existing
reservation. If the reservation that would be created for the null
derived capability is found to overlap with any existing reservation,
the syscall fails with a -ERESERVATION error code.

According to the spec, the lifetime of reservations is generally
undefined. While the reservation is eventually destroyed and the address
range is made available, there should be no assumption as to when that
will happen. But in the current implementation, the reservation is
destroyed as soon as the last mapping is unmapped. Here a test is added
to ensure that mmap(owning_cap, ...MAP_FIXED) fails because the
underlying reservation is destroyed synchronously.

A partial unmap within a particular reservation still allows the rest of
the region to be accessible. Tests to verify the same have been added.

Signed-off-by: Chaitanya S Prakash <chaitanyas.prakash@arm.com>