io_uring: Add explicit capability checks
When working in PCuABI, user space pointers are capabilities that carry
metadata such as bounds and read/write permissions. These two are
automatically checked when direct accesses are performed.

The io_uring interface provides a facility to perform asynchronous I/O
operations, and does not always access the data directly, but it relies
on recalling the pages where it's held from the virtual memory. In order
to this, it extracts the address from the capability ignoring the
metadata.

This commit adds explicit capability checks whenever userspace memory
regions pointed by capabilities are accessed indirectly, as
specified in the io_uring section of the PCuABI specification [1].

[1] https://git.morello-project.org/morello/kernel/linux/-/wikis/Morello-pure-capability-kernel-user-Linux-ABI-specification

Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>