mm/gup: Create user pointer when calling access_ok()
GUP does not get passed an actual user pointer, just a user address.
In PCuABI, it is up to the caller to check the user pointer's
capability metadata, if appropriate. access_ok() however expects a
user pointer, so create an invalid one to make it happy - this is
sufficient as access_ok() only cares about the address.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>