mm/mincore: Temporarily create a user pointer for access_ok()
access_ok() expects a user pointer, but the mincore handler takes an
unsigned long at the moment. For the time being, use as_user_ptr()
to create an artificial user pointer and make access_ok() happy.
Eventually this is to be reverted as the mincore syscall should take
a capability in PCuABI, as per the specification.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>