selftests: vDSO: fix Makefile so that it uses lib.mk correctly
kselftest.rst states that flags must be specified before including lib.mk,
but the vDSO selftest Makefile does not follow this order. As a result,
changes made by lib.mk to flags and other variables are overwritten by the
Makefile. For example, it is impossible to pass CFLAGS to the compiler via
make.

Rectify this by including lib.mk after assigning flag values.

Also change the paths of the generated programs from absolute to relative
paths as lib.mk will now correctly prepend the output directory path to
the program name as intended.

Signed-off-by: Aditya Deshpande <aditya.deshpande@arm.com>