fs/binfmt_elf: Improve SP manipulation in PCuABI
In PCuABI, we should derive capabilities from the user root
capability as rarely as possible. The ELF loader currently does this
many times in create_elf_tables(), as it represents positions on the
stack as simple addresses in many cases.

This patch improves the situation by deriving all pointers to the
stack from a single stack pointer variable (sp). As a result
uaddr_to_user_ptr_safe() is now only used once to initialise sp.

The type of sp is changed to void __user * as it does not actually
point to aligned data at the beginning of the function
(STACK_ALLOC() allocates a string on the stack without any
particular alignment). Similarly, the type of the u_* variables is
changed to char __user * as they point to these (unaligned) strings,
not arrays of elf_addr_t. Once sp has been aligned using
STACK_ROUND(), a new pointer stack_item, of the appropriate type, is
used to write the pointer-sized "stack items".

This patch should not induce any user-visible change.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>