usb: core: Fix copy of URB from userspace
When working with PCuABI, user pointers carry metadata as they are
capabilities and they have a tag associated to them, which is placed in
another region of memory. The kernel's copy_from_user function does not
support them, and as a consequence their tag is not copied in the
process.

This commit changes the used function to copy_from_user_with_ptr which
exists to solve this problem.

Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>