Documentation: cheri: Add explicit checking to porting guide
Now that explicit capability checking is carried out in the core
kernel and the drivers that are already supported in PCuABI, add a
section to the porting guide explaining in which situations explicit
checking might be required and what should be done.

Also add a few notes to the table of code examples, to make it clear
that using the new user pointer API is not necessarily enough in
itself.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>