linux/user_ptr.h: Introduce user_ptr_{base,limit}
Add generic helpers to obtain the bounds (base and limit) of a user
pointer. These are useful when an algorithm wants to know whether it
is possible to access memory beyond the bounds of a given object.
There should be no need for them for checking in-bound accesses,
as those checks are already carried out during uaccess, or
explicitly via check_user_ptr_*() (which perform a full validity
check, not just a bounds check).

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>