selftests/arm64: morello: Improve __EXPECT()
__EXPECT() currently prints the stringified arguments on failure but
not their value, which is not very helpful.

Let's use the same message as the original macro in
kselftest_harness.h, printing the values too. For simplicity's sake
they are always printed as signed, which is generally appropriate
(only 64-bit unsigned values with the top bit set are an issue, and
these are rarely encountered).

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>