arm64: uaccess: Move macros from futex.h to uaccess.h
In preparation of enabling capability-based accesses in
get_user/put_user, move the PCuABI-related macros from futex.h to
uaccess.h (the former includes the latter). Also give the A64/C64
macros a more generic name, which is going to be useful in uaccess.h
to support kernel and user accesses in the same macro.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>