arm64: uaccess: Switch to capability-based {get,put}_user in PCuABI
In PCuABI, {get,put}_user currently access user memory via standard
64-bit pointers. This patch switches to capability-based accesses:
instead of extracting the address of the input capability, it is
directly used to make the access. As a result, {get,put}_user
validate the capability metadata, failing with -EFAULT if the
hardware checks fail (causing a capability exception).

Unprivileged load and store instructions operating on a capability
base register (e.g. ldtr/sttr x0, [c1]) are only available in the
C64 ISA, we therefore need to switch to C64 before executing them
(and then return to A64). We also need to use the "C" asm constraint
for capability registers. This requirement has recently disappeared
in Clang, but for backwards-compatibility we stick to using that
constraint for now.

Because __{get,put}_mem_asm can operate on either a kernel or user
pointer, fallback macros are defined for the kernel case. The
Morello-specific variants that load or store a capability
(__morello_raw_{get,put}_user_cap) are only passed user pointers, so
the user macros can directly be used there.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>