arm64: lib: Simplify copy_*_user register allocation
copy_template currently copies the destination pointer to x6 then
operates on that copy, while it directly operates on the source
pointer (x1). In both cases a copy of the original pointer is
required for the final fixup in copy_*_user, but not in
copy_template itself.

Make things a little easier to follow by saving both pointers
in copy_*_user, letting copy_template operate on the original
pointers (x0 and x1) directly.

While at it, remove the "Returns" comment in copy_template, which is
irrelevant.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>