arm64: lib: Switch to capability-based copy_*_user in PCuABI
We have already amended {get,put}_user so that they access user
memory directly via the user capability in PCuABI; now is time to
convert copy_*_user too. As a result, the copy will be aborted if
the user capability is unsuitable to perform the access (potentially
partway, in case the tail of the targeted region is out of bounds).

__arch_copy_{from,to}_user present an additional challenge in
that they are implemented fully in assembly. Fortunately, the
registers holding the source and destination pointers are mostly
used as base registers for load/store instructions. After switching
to C64, such instructions operate on C registers instead of X, so it
becomes simply a matter of modifying the register aliases in PCuABI;
the req_reg_pcuabi macro is introduced for that purpose. Explicit
ADD instructions are also used in the user_{ldst,ldp,stp} helpers;
those are unproblematic as they can operate on both X and C
registers.

In the few situations where the pointers are being somehow
inspected, we keep operating on their address only, by using the
corresponding X register; srcx and dstx are introduced for that
purpose. This is necessary in some cases due to the instruction
simply not accepting C registers (e.g. TST), while in others it is
rather a matter of convenience, as it means we don't need to convert
additional register aliases to capabilities (CMP, SUB).

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>