kselftests/arm64: morello: mmap: Test unrepresentable addresses
Morello uses a compressed capability format which makes it difficult to
represent bounds with arbitrary precision. As the corresponding address
range of a given memory mapping may not be exactly representable as valid
capability bounds, a test to verify that the PCuABI kernel is able to
mmap/munmap those addresses has been added.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>
Co-Developed-by: Chaitanya S Prakash <ChaitanyaS.Prakash@arm.com>
Signed-off-by: Chaitanya S Prakash <ChaitanyaS.Prakash@arm.com>