linux/types.h: Redefine uintptr_t to __uintcap_t under purecap
This change is made in preparation for the introduction of a purecap
vDSO. While the vDSO is userspace shared library, it does use some
kernel headers. Therefore, under purecap, any pointers in those headers
must be compiled to capabilites.

Note that this does not make any part of the actual kernel purecap.

Signed-off-by: Aditya Deshpande <aditya.deshpande@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>