bpf: debug: bpf_jit_enable=0 by default
While messing with the JIT, it's useful to have it turned off by
default; since systemd runs some bpf on startup, if we break the JIT
we'll end up not being able to boot.

note: CONFIG_BPF_JIT_DEFAULT_ON is automatically selected by
ARCH_WANT_DEFAULT_BPF_JIT and cannot be selected as part of a
defconfig, hence the approach here.

Once booted, turn on the JIT via:
> echo "2" > /proc/sys/net/core/bpf_jit_enable

Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>