io_uring: Use full pointers for IORING_SETUP_NO_MMAP
Commit 03d89a2de25b ("io_uring: support for user allocated memory
for rings/sqes") introduced struct io_{s,c}qring_offsets::user_addr
to allow the user to specify ring mappings themselves. As usual,
these fields are typed as __u64, although they are meant to hold
pointers. This means that only the mapping address can be passed,
and not a full pointer in PCuABI.

Move to operating on full pointers and perform the usual
adaptations:

* __u64 becomes __kernel_uintptr_t (able to hold a full pointer in
  any ABI)

* Propagate the value as a user pointer down to the point of use
  (__io_uaddr_map()).

* To keep the ABI unchanged in compat64, introduce compat handling
  for io_uring_setup(). struct io_uring_params is both copied in and
  out, so we need to convert the layout in both directions. Not all
  fields are necessarily relevant in both cases, but for
  simplicity's sake and forward-compatibility, all of them are
  copied.

Finally, because the user memory is accessed via a kernel mapping,
the provided pointer needs to be explicitly checked before calling
pin_user_pages_fast(). Here again, for simplicity's sake, we require
a RW pointer in both cases. Write access may not be strictly
required for struct io_sqring_offsets::user_addr, but userspace
needs to have a writeable mapping / pointer to be able to make use
of it at all, so this should not make much difference in practice.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>