io_uring/cmd: Convert compat64 pointers for OP_{G,S}ETSOCKOPT
struct io_uring_sqe::optval is a strange field, because in reality
it is part of the cmd payload (for IORING_OP_URING_CMD). The layout
of this payload is not converted before the SQE is submitted (in
convert_compat64_io_uring_sqe()), because it is specific to
each .uring_cmd handler. Even if we wanted to handle all the
conversions upfront, it would be difficult to achieve as the target
file (and thus the .uring_cmd handler) has not been determined at
that point.

Since optval is a user pointer, we need to perform the usual
conversion for compat64. This is done by getting a pointer to
the payload using io_uring_sqe_cmd(), and then interpreting that
payload as either __u64 or __kernel_uintptr_t. The optval field is
no longer used, as it is not meaningful in compat64. Since struct
io_uring_sqe is uapi, the field is retained for userspace usage.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>