mm/reserv: Add address space reservation API
This patch introduces helpers to manage address space reservations
and owning capabilities, as introduced in the PCuABI specification.
This interface prevents two unrelated owning capabilities created by
the kernel from overlapping.

The reservation interface stores virtual address ranges as
reservation entries, which is the same as the bounds of the owning
capability provided by the kernel to userspace. It also stores the
owning capability permissions to manage the future syscall requests
for updating permissions.

The reservation interface follows a few basic rules:

- Reservations can only be created or destroyed but never expanded or
  shrunk. Reservations are created when new memory mapping is made outside
  of an existing reservation.
- A single reservation can have many mappings. However, unused regions
  of the reservation cannot be reused again.
- The Reservation start address is aligned to CHERI representable base.
- The Reservation length value is aligned to CHERI representable length.

More rules about the address space reservation interface can be found in
the PCuABI specification.

Signed-off-by: Amit Daniel Kachhap <amitdaniel.kachhap@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>