arm64: morello: Context-switch RCSP at EL1 too
RCSP is currently considered as an EL0 register. However, Restricted
registers are not banked, and as a result this assumes that the
kernel itself does not modify it.

To enable the kernel to make use of RCSP, save and restore it when
taking an exception from / returning to EL1, in addition to EL0.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>