arm64: morello: Context-switch RDDC on kernel entry/return
RDDC is currently considered as an EL0 register. However, Restricted
registers are not banked, and as a result this assumes that the
kernel itself does not modify it.

To enable the kernel to make use of RDDC, context-switch it on every
kernel entry/return, instead of when rescheduling (__switch_to). As
a result its value is saved in pt_regs instead of morello_state, and
a few places need fixing up accordingly.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>