arm64: Switch to standard asm constraints for capabilities
Morello Clang used to support specifying capability variables in
inline assembly only via the "C" constraint. As of release 1.7 of
Morello LLVM, standard constraints ("r", "Q") are also supported,
like in Morello GCC.

We can therefore simplify various pieces of inline assembly by
restoring the original constraints and getting rid of the
ABI-specific macros. Inline assembly that always operates on
capabilities is left unchanged, as the "C" constraint remains
supported and it is more explicit (only a capability may be provided
in that case).

This commit partially reverts a number of uaccess-related commits,
as well as a vDSO-related commit.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>